third_party/blink/web_tests/external/wpt/css/css-sizing/aspect-ratio/replaced-element-043-expected.txt

This is a testharness.js-based test.
Found 35 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] canvas 15
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px" data-expected-width="20" data-expected-height="15"></canvas>\nheight expected 15 but got 10
[FAIL] canvas 21
  assert_equals: \n<canvas width="15" height="15" style="min-height: 10px" data-expected-width="20" data-expected-height="15"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 22
  assert_equals: \n<canvas width="15" height="15" style="min-height: 15px" data-expected-width="30" data-expected-height="15"></canvas>\nwidth expected 30 but got 15
[FAIL] canvas 27
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px" data-expected-width="15" data-expected-height="10"></canvas>\nheight expected 10 but got 15
[FAIL] canvas 33
  assert_equals: \n<canvas width="15" height="15" style="max-height: 10px" data-expected-width="15" data-expected-height="10"></canvas>\nwidth expected 15 but got 20
[FAIL] canvas 39
  assert_equals: \n<canvas width="15" height="15" style="min-width: 0px; min-height: 10px" data-expected-width="20" data-expected-height="15"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 40
  assert_equals: \n<canvas width="15" height="15" style="min-width: 0px; min-height: 15px" data-expected-width="30" data-expected-height="15"></canvas>\nwidth expected 30 but got 15
[FAIL] canvas 45
  assert_equals: \n<canvas width="15" height="15" style="min-width: 10px; min-height: 10px" data-expected-width="20" data-expected-height="15"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 46
  assert_equals: \n<canvas width="15" height="15" style="min-width: 10px; min-height: 15px" data-expected-width="30" data-expected-height="15"></canvas>\nwidth expected 30 but got 15
[FAIL] canvas 49
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px; min-height: 0px" data-expected-width="20" data-expected-height="15"></canvas>\nheight expected 15 but got 10
[FAIL] canvas 50
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px; min-height: 5px" data-expected-width="20" data-expected-height="15"></canvas>\nheight expected 15 but got 10
[FAIL] canvas 51
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px; min-height: 10px" data-expected-width="20" data-expected-height="15"></canvas>\nheight expected 15 but got 10
[FAIL] canvas 52
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px; min-height: 15px" data-expected-width="30" data-expected-height="15"></canvas>\nwidth expected 30 but got 20
[FAIL] canvas 75
  assert_equals: \n<canvas width="15" height="15" style="min-width: 0px; max-height: 10px" data-expected-width="15" data-expected-height="10"></canvas>\nwidth expected 15 but got 20
[FAIL] canvas 81
  assert_equals: \n<canvas width="15" height="15" style="min-width: 10px; max-height: 10px" data-expected-width="15" data-expected-height="10"></canvas>\nwidth expected 15 but got 20
[FAIL] canvas 88
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px; max-height: 15px" data-expected-width="20" data-expected-height="15"></canvas>\nheight expected 15 but got 10
[FAIL] canvas 89
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px; max-height: 20px" data-expected-width="20" data-expected-height="15"></canvas>\nheight expected 15 but got 10
[FAIL] canvas 90
  assert_equals: \n<canvas width="15" height="15" style="min-width: 20px; max-height: 25px" data-expected-width="20" data-expected-height="15"></canvas>\nheight expected 15 but got 10
[FAIL] canvas 121
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; min-height: 0px" data-expected-width="15" data-expected-height="10"></canvas>\nheight expected 10 but got 15
[FAIL] canvas 122
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; min-height: 5px" data-expected-width="15" data-expected-height="10"></canvas>\nheight expected 10 but got 15
[FAIL] canvas 123
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; min-height: 10px" data-expected-width="20" data-expected-height="10"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 124
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; min-height: 15px" data-expected-width="20" data-expected-height="15"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 129
  assert_equals: \n<canvas width="15" height="15" style="max-width: 30px; min-height: 10px" data-expected-width="20" data-expected-height="15"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 130
  assert_equals: \n<canvas width="15" height="15" style="max-width: 30px; min-height: 15px" data-expected-width="30" data-expected-height="15"></canvas>\nwidth expected 30 but got 15
[FAIL] canvas 135
  assert_equals: \n<canvas width="15" height="15" style="max-width: 40px; min-height: 10px" data-expected-width="20" data-expected-height="15"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 136
  assert_equals: \n<canvas width="15" height="15" style="max-width: 40px; min-height: 15px" data-expected-width="30" data-expected-height="15"></canvas>\nwidth expected 30 but got 15
[FAIL] canvas 141
  assert_equals: \n<canvas width="15" height="15" style="max-width: 50px; min-height: 10px" data-expected-width="20" data-expected-height="15"></canvas>\nwidth expected 20 but got 15
[FAIL] canvas 142
  assert_equals: \n<canvas width="15" height="15" style="max-width: 50px; min-height: 15px" data-expected-width="30" data-expected-height="15"></canvas>\nwidth expected 30 but got 15
[FAIL] canvas 159
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; max-height: 10px" data-expected-width="15" data-expected-height="10"></canvas>\nwidth expected 15 but got 20
[FAIL] canvas 160
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; max-height: 15px" data-expected-width="15" data-expected-height="10"></canvas>\nheight expected 10 but got 15
[FAIL] canvas 161
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; max-height: 20px" data-expected-width="15" data-expected-height="10"></canvas>\nheight expected 10 but got 15
[FAIL] canvas 162
  assert_equals: \n<canvas width="15" height="15" style="max-width: 20px; max-height: 25px" data-expected-width="15" data-expected-height="10"></canvas>\nheight expected 10 but got 15
[FAIL] canvas 165
  assert_equals: \n<canvas width="15" height="15" style="max-width: 30px; max-height: 10px" data-expected-width="15" data-expected-height="10"></canvas>\nwidth expected 15 but got 20
[FAIL] canvas 171
  assert_equals: \n<canvas width="15" height="15" style="max-width: 40px; max-height: 10px" data-expected-width="15" data-expected-height="10"></canvas>\nwidth expected 15 but got 20
[FAIL] canvas 177
  assert_equals: \n<canvas width="15" height="15" style="max-width: 50px; max-height: 10px" data-expected-width="15" data-expected-height="10"></canvas>\nwidth expected 15 but got 20
Harness: the test ran to completion.
